proyecto_1_etl_pipeline/requirements.txt
pandas
python-dotenv
matplotlib